Move new docs pages to MDX files
