+++ contents/docs/data.mdx
@@ -1,0 +1,15 @@
+---
+title: Data
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import DataPage from '../../src/components/Docs/Pages/Data'
+
+<DataPage />
