+++ contents/docs/experiments.mdx
@@ -1,0 +1,15 @@
+---
+title: Experiments
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import ExperimentsPage from '../../src/components/Docs/Pages/Experiments'
+
+<ExperimentsPage />
