+++ contents/docs/feature-flags.mdx
@@ -1,0 +1,15 @@
+---
+title: Feature flags
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import FeatureFlagsPage from '../../src/components/Docs/Pages/FeatureFlags'
+
+<FeatureFlagsPage />
