+++ contents/docs/index.mdx
@@ -1,0 +1,15 @@
+---
+title: Docs
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import DocsPage from '../../src/components/Docs/Pages'
+
+<DocsPage />
