+++ contents/docs/integrations.mdx
@@ -1,0 +1,15 @@
+---
+title: Integrations
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import IntegrationsPage from '../../src/components/Docs/Pages/Integrations'
+
+<IntegrationsPage />
