+++ contents/docs/product-analytics.mdx
@@ -1,0 +1,15 @@
+---
+title: Product analytics
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import ProductAnalyticsPage from '../../src/components/Docs/Pages/ProductAnalytics'
+
+<ProductAnalyticsPage />
