+++ contents/docs/session-replay.mdx
@@ -1,0 +1,15 @@
+---
+title: Session replay
+hideAnchor: true
+isArticle: false
+hideSurvey: true
+hideQuestions: true
+hideNextPost: true
+hideTitle: true
+hideLastUpdated: true
+hideBreadcrumb: true
+---
+
+import SessionReplayPage from '../../src/components/Docs/Pages/SessionReplay'
+
+<SessionReplayPage />
