+++ gatsby-browser.tsx
@@ -40,16 +40,7 @@ export const wrapPageElement = ({ element, props }) => {
     return props.custom404 || !props.data ? (
         element
     ) : /^handbook|^docs\/(?!api)|^manual/.test(slug) &&
-      ![
-          'docs/api/post-only-endpoints',
-          'docs/api/user',
-          'docs/integrations',
-          'docs/product-analytics',
-          'docs/session-replay',
-          'docs/feature-flags',
-          'docs/experiments',
-          'docs/data',
-      ].includes(slug) ? (
+      !['docs/api/post-only-endpoints', 'docs/api/user'].includes(slug) ? (
         <HandbookLayout {...props} />
     ) : /^session-replay|^product-analytics|^feature-flags|^ab-testing|^product-os/.test(slug) ? (
         <Product {...props} />
