+++ src/mdxGlobalComponents.js
@@ -6,6 +6,7 @@ import { BorderWrapper } from './components/BorderWrapper'
 import { CallToAction } from './components/CallToAction'
 import { Caption } from './components/Caption'
 import { CompensationCalculator } from './components/CompensationCalculator'
+import { Pages } from './components/Docs/Pages'
 import { FeatureAvailability } from './components/FeatureAvailability'
 import { GDPRForm } from './components/GDPRForm'
 import { HiddenSection } from './components/HiddenSection'
@@ -25,6 +26,7 @@ export const shortcodes = {
     CallToAction,
     Caption,
     CompensationCalculator,
+    Pages,
     FeatureAvailability,
     GDPRForm,
     HiddenSection,
